Update the TXN_SPIKE behaviour to match new specifications.
The TXN_SPIKE event now just switch all the permissible chromatin to the TRANSCRIBABLE state, without any change in the rates. This means it has to be paired with an ACTIVATE event. Please note that if ACTIVATE and TXN_SPIKE are scheduled at the same time, they are commutative, i.e. it does not matter which one is executed first.

Closes #19