Fix building with icc
For Release build type, the build process was requiring a profile generated with PGO to be present. Now it is possible to build Release without any PGO, if desired.
The variable ENABLE_PGO_USE can be passed to Cmake (and can be set in cmake-with-icc.sh) to switch PGO usage ON/OFF.