Enable local simulation of simplified pahse behaviors
I added an option in the Grid Intializer to make the lattice boundary inactive RBP, so as to not let the chromatin stick to it. Also, I made a script to locally run example simulations that show the phase behavior in a simpler manner.mSo one can simulate the three main cases: inactive, fully active (microphase separation), and inhibited after activity (phase separation).